Write apps bundle to source directory
At runtime these need to be extracted from an asset so that the python
code can read them. We could write our own code to extract the files
from the AssetManager at runtime. However, since we've already told
chaquopy to do that for the testapp package, just put them in there.