Reload zipfile module
The template-ui zip files are compressed with LZMA and need to use that
support from the `zipfile` module. Unfortunately, chaquopy imports
`zipfile` while bootstrapping the runtime to unpack its own assets but
doesn't include the `_lzma.so` module in the bootstrap component. Reload
`zipfile` at the end of the setup process so it has access to that
module.